+++ src/types/usageRequest.ts
@@ -3,7 +3,7 @@ import { UsageRequestDetail } from "./usageRequestDetail";
 import { UsageRequestMessage } from "./usageRequestMessage";
 
 export type UsageRequest = {
-  id: string;
+  request_id: string;
   created: string;
   path: string;
   accessor: string;